+++ background/service-worker.js
@@ -8,6 +8,8 @@ importScripts(
     '../utils/storage-manager.sync.js',
     '../utils/storage-manager.presets.js',
     '../utils/storage-manager.chatmap.js',
+    '../utils/storage-manager.local.js',
+    '../utils/storage-manager.init.js',
     '../utils/storage-manager.syncnow.js',
     '../utils/storage-manager.js'
 );